add binding adapter, removes some reductant codes
