+++ app/src/main/java/com/example/notes/activities/utils/BindingAdapter.kt
@@ -1,0 +1,15 @@
+package com.example.notes.activities.utils
+
+import android.graphics.BitmapFactory
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+
+@BindingAdapter("setBitmapSource")
+fun ImageView.setBitmapSource(pathname: String?) {
+    if (pathname !== null) {
+        this.setImageBitmap(BitmapFactory.decodeFile(pathname))
+        this.visibility = ImageView.VISIBLE
+    } else {
+        this.visibility = ImageView.GONE
+    }
+}
