+++ app/src/main/res/layout/list_item.xml
@@ -28,6 +28,7 @@
             android:orientation="vertical">
 
             <ImageView
+                setBitmapSource="@{noteEntity.imagePath}"
                 android:id="@+id/item_cover"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
